MER-2859 Move VrUserAgents to its own module
